refactor(context): migrate assembler_helpers to zeph-agent-context
Move all helper functions from `zeph-core/agent/context/assembler_helpers.rs`
into `zeph-agent-context/src/helpers.rs`, deleting the original file.

- Delete `assembler_helpers.rs` from `zeph-core` (547 lines removed)
- Migrate helpers to `zeph-agent-context::helpers` with `ContextAssemblyView`
  parameter replacing `MemoryState`; `ContextError` replacing `AgentError`
- Add PERSONA_PREFIX, TRAJECTORY_PREFIX, TREE_MEMORY_PREFIX, REASONING_PREFIX
  constants to helpers; remove duplicates from `context/mod.rs`
- Remove `#[cfg(test)] use zeph_agent_context::helpers as assembler_helpers`
  re-export alias from `context/mod.rs`; re-export the four prefix constants
  via `pub(super) use zeph_agent_context::helpers::{...}` instead
- Update call sites in `assembly.rs` and tests to use `zeph_agent_context::helpers`
  directly

Closes the backwards-compat shim left after the scaffold stage.
