+++ crates/zeph-core/src/agent/context/mod.rs
@@ -1,8 +1,6 @@
 // SPDX-FileCopyrightText: 2026 Andrei G <bug-ops>
 // SPDX-License-Identifier: MIT OR Apache-2.0
 
-#[cfg(test)]
-pub(super) mod assembler_helpers;
 mod assembly;
 #[cfg(feature = "self-check")]
 pub(crate) mod retrieved;
@@ -11,6 +9,9 @@ mod summarization;
 use super::{Agent, Channel};
 
 pub(super) use crate::text::truncate_to_chars as truncate_chars;
+pub(super) use zeph_agent_context::helpers::{
+    PERSONA_PREFIX, REASONING_PREFIX, TRAJECTORY_PREFIX, TREE_MEMORY_PREFIX,
+};
 pub(super) use zeph_context::slot::{cap_summary, chunk_messages};
 
 /// Return type from `compact_context()` that distinguishes between successful compaction,
@@ -31,11 +32,6 @@ pub(super) enum CompactionOutcome {
     /// No compaction was performed (too few messages, empty `to_compact`, etc.).
     NoChange,
 }
-
-pub(super) const PERSONA_PREFIX: &str = "[Persona context]\n";
-pub(super) const TRAJECTORY_PREFIX: &str = "[Past experience]\n";
-pub(super) const TREE_MEMORY_PREFIX: &str = "[Memory summary]\n";
-pub(super) const REASONING_PREFIX: &str = "[Reasoning Strategy]\n";
 
 impl<C: Channel> Agent<C> {
     pub(super) fn compaction_tier(&self) -> super::context_manager::CompactionTier {
